Add casts for the float lerp() operator
This avoids the warning from MSVC:

  warning C4244: 'return': conversion from 'double' to 'float', possible loss of data

And it's similar to what we do for the integer lerp().

+++ src/graphene-private.h
@@ -111,7 +111,7 @@ graphene_approx_val (float a, float b)
 static inline float
 graphene_flerpf (float a, float b, double factor)
 {
-  return ((1.f - factor) * a) + (factor * b);
+  return (float) ((1.0 - factor) * (double) a) + (factor * (double) b);
 }
 
 static inline double
